Refactored /api/users/user/[id]/route.js to getUser server function.
