+++ src/app/user/[id]/posts/page.js
@@ -1,12 +1,12 @@
 'use client'
 
 import { countUserComments, countUserPosts } from '@/lib/actions/utils'
-import { getUser } from '@/lib/db'
 import PostsPage from '@/app/posts/page'
 import { Loader } from '@/ui/loaders/Loader'
 import { UserProfileShimmer } from '@/ui/loaders/UserProfileShimmer'
 import { ProfilePreview } from '@/ui/profile/ProfilePreview'
 import { useEffect, useState } from 'react'
+import { getUser } from '@/lib/actions/user'
 
 export default function UserProfile({ params, searchParams }) {
   const [userData, setUserData] = useState(null)
