+++ src/lib/actions/auth.js
@@ -6,17 +6,23 @@ import User from '../models/User'
 
 export async function getUserByEmail(email) {
   if (!validator.isEmail(email)) {
-    throw new Error('Invalid email')
+    console.error('Invalid email')
+    return null
   }
 
-  await connectToDatabase()
-  const user = await User.findOne({ email: email })
+  try {
+    await connectToDatabase()
+    const user = await User.findOne({ email: email })
 
-  if (!user) return null
+    if (!user) return null
 
-  return {
-    password: user.password,
-    is_active: user.is_active,
-    status: user.status,
+    return {
+      password: user.password,
+      is_active: user.is_active,
+      status: user.status,
+    }
+  } catch (error) {
+    console.error('Error in getUserByEmail, error:', error)
+    return null
   }
 }
