+++ src/lib/actions/favorites.js
@@ -4,8 +4,8 @@ import { authOptions } from '@/app/api/auth/[...nextauth]/authOptions'
 import { getServerSession } from 'next-auth/next'
 import { redirect } from 'next/navigation'
 import { isUUID } from 'validator'
-import { getUser } from '../db'
 import User from '../models/User'
+import { getUser } from './user'
 import allowedPostIds from '../security/allowedPostIds'
 import { returnToast, setToast, toast } from '../toasts/ToastUtils'
